+++ src/client.rs
@@ -1,9 +1,10 @@
 use crate::configuration::ClientConfiguration;
-use crate::interfaces::setup_client_interfaces;
+use crate::interfaces::{setup_client_interfaces, NANOSECOND, SECOND};
 use crate::notifications::{ClientReadyMessage, Key, Notification};
 use crate::shutdown::listen_for_shutdown;
-use crate::{Error, LIB_LOG_TARGET};
-use log::{debug, info, warn};
+use crate::{Error, CHANNEL_BUFFER, LIB_LOG_TARGET};
+use log::{debug, error, info, trace, warn};
+use std::sync::{Arc, Mutex};
 use tokio::sync::{broadcast, mpsc, watch};
 
 const DEFAULT_WAIT_FOR_SHUTDOWN_SECS: u64 = 2;
@@ -17,12 +18,12 @@ pub async fn start_client(
     wait_for_shutdown_secs: Option<u64>,
 ) -> Result<(), Error> {
     let (shutdown_tx, shutdown_rx) = watch::channel(false);
-    let (interface_tx, _interface_rx) = broadcast::channel(100);
+    let (interface_tx, interface_rx) = broadcast::channel(CHANNEL_BUFFER);
     let key = client_config.key().clone();
 
     // Setup interfaces to send notifications to
     let interfaces = client_config.interfaces();
-    setup_client_interfaces(interfaces, interface_tx.subscribe(), shutdown_rx.clone()).await?;
+    setup_client_interfaces(interfaces, interface_rx, shutdown_rx.clone()).await?;
 
     // Monitor for incoming notifications
     tokio::spawn(async move {
@@ -30,12 +31,34 @@ pub async fn start_client(
     });
 
     // Shutdown
-    let shutdown_secs = match wait_for_shutdown_secs {
-        None => DEFAULT_WAIT_FOR_SHUTDOWN_SECS,
-        Some(secs) => secs,
-    };
-    info!(target: LIB_LOG_TARGET, "Listening for shutdown signals");
-    listen_for_shutdown(shutdown_tx, shutdown_secs).await;
+    listen_for_shutdown(shutdown_tx, get_shutdown_wait_time(wait_for_shutdown_secs)).await;
+
+    Ok(())
+}
+
+/// Start the client with provided [`ClientConfiguration`] and `Arc<Mutex<Vec<ClientReadyMessage>>>`.
+///
+/// Client listens for shutdown signals SIGTERM & SIGINT on Unix or CTRL-BREAK and CTRL-C on Windows.
+pub async fn start_client_arc(
+    client_config: ClientConfiguration,
+    notifications: Arc<Mutex<Vec<ClientReadyMessage>>>,
+    wait_for_shutdown_secs: Option<u64>,
+) -> Result<(), Error> {
+    let (shutdown_tx, shutdown_rx) = watch::channel(false);
+    let (interface_tx, interface_rx) = broadcast::channel(CHANNEL_BUFFER);
+    let key = client_config.key().clone();
+
+    // Setup interfaces to send notifications to
+    let interfaces = client_config.interfaces();
+    setup_client_interfaces(interfaces, interface_rx, shutdown_rx.clone()).await?;
+
+    // Monitor for incoming notifications
+    tokio::spawn(async move {
+        receive_notifications_arc(notifications, interface_tx, shutdown_rx.clone(), key).await;
+    });
+
+    // Shutdown
+    listen_for_shutdown(shutdown_tx, get_shutdown_wait_time(wait_for_shutdown_secs)).await;
 
     Ok(())
 }
@@ -57,14 +80,68 @@ async fn receive_notifications(
                     debug!(target: LIB_LOG_TARGET, "Client Sending Notification: {:?}", notification);
                     match interface_tx.send(notification) {
                         Ok(ok) => debug!(target: LIB_LOG_TARGET, "Message passed to client {} interfaces", ok),
-                        Err(error) => warn!(target: LIB_LOG_TARGET, "Client broadcast channel send error: {}", error),
+                        Err(error) => {
+                            error!(target: LIB_LOG_TARGET, "Client broadcast channel send error: {}", error);
+                            break;
+                        },
                     }
                 }
             }
 
             _ = shutdown_rx.changed() => {
+                trace!(target: LIB_LOG_TARGET, "Shutdown receive_notifications");
                  break;
                 }
+
+            _ = tokio::time::sleep(SECOND) => {
+                trace!(target: LIB_LOG_TARGET, "Sleep timeout reached for receive_notifications");
+            }
         }
+        tokio::time::sleep(NANOSECOND).await;
+    }
+}
+
+async fn receive_notifications_arc(
+    notifications: Arc<Mutex<Vec<ClientReadyMessage>>>,
+    interface_tx: broadcast::Sender<Notification>,
+    shutdown: watch::Receiver<bool>,
+    key: Key,
+) {
+    info!(target: LIB_LOG_TARGET, "Client waiting for notifications");
+
+    let mut shutdown_rx = shutdown.clone();
+    loop {
+        tokio::select! {
+            _ = shutdown_rx.changed() => {
+                trace!(target: LIB_LOG_TARGET, "Shutdown receive_notifications_arc");
+                 break;
+                }
+
+            _ = tokio::time::sleep(SECOND) => {
+                trace!(target: LIB_LOG_TARGET, "Sleep timeout reached for receive_notifications_arc");
+            }
+        }
+
+        let messages: Vec<ClientReadyMessage> = notifications.lock().unwrap().drain(0..).collect();
+
+        if !messages.is_empty() {
+            for client_ready_msg in messages {
+                let notification = client_ready_msg.to_notification(&key);
+                debug!(target: LIB_LOG_TARGET, "Client attempting to send Notification: {:?}", notification);
+
+                match interface_tx.send(notification) {
+                    Ok(ok) => debug!(target: LIB_LOG_TARGET, "Message passed to client interfaces: {}", ok),
+                    Err(error) => warn!(target: LIB_LOG_TARGET, "Client broadcast channel send error: {}", error),
+                }
+            }
+        }
+        tokio::time::sleep(NANOSECOND).await;
+    }
+}
+
+fn get_shutdown_wait_time(seconds: Option<u64>) -> u64 {
+    match seconds {
+        None => DEFAULT_WAIT_FOR_SHUTDOWN_SECS,
+        Some(secs) => secs,
     }
 }