+++ src/lib.rs
@@ -82,7 +82,7 @@ mod server;
 pub(crate) mod shutdown;
 
 #[cfg(feature = "client")]
-pub use self::client::start_client;
+pub use self::client::{start_client, start_client_arc};
 #[cfg(feature = "client")]
 pub use self::configuration::ClientConfiguration;
 #[cfg(feature = "server")]
@@ -93,3 +93,5 @@ pub use self::server::start_server;
 
 /// Logging target value used for the library.
 pub const LIB_LOG_TARGET: &str = "pass_it_on";
+#[allow(dead_code)]
+const CHANNEL_BUFFER: usize = 200;