+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v0.4.0
+## Features
+- Added `start_client_arc` which has the same functionality of the `start_client` function but accepts a `Arc<Mutex<Vec<ClientReadyMessage>>>`.
+
+## Fixes
+- Internal changes to client to ensure it continues to receive passed notifications.
+- Fix cfg statement for pipe interface
+
 # v0.3.3
 ## Fixes
 - Ensure client shuts down correctly